+++ .github/workflows/pull-request.yml
@@ -1,4 +1,4 @@
-name: Create and publish a Docker image
+name: Run Angular Tests on PR
 
 on:
   pull_request:
