+++ angular/whereami/src/app/views/create-game/drawing-manager/drawing-manager.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { GOOGLE_MAP_STUB, GOOGLE_TESTING_PROVIDER } from 'src/app/Test/web-socket-service-spy.spec';
+import { GOOGLE_MAP_STUB, GOOGLE_TESTING_PROVIDER } from 'src/app/Test/testutils.spec';
 
 import { DrawingManagerComponent } from './drawing-manager.component';
